11.5-attention/requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.61
botocore==1.17.61
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
docutils==0.15.2
gast==0.2.2
gensim==3.8.3
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.6
Janome==0.4.0
jmespath==0.10.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
nltk==3.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.1.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pygobject==3.26.1
python-apt==1.6.4
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.25
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
SecretStorage==2.3.1
six==1.13.0
sklearn==0.0
smart-open==2.1.1
soupsieve==2.0.1
tensorboard==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.49.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
